Create problem3.py
The Python program reads a given text ﬁle and counts the positional frequencies of each letter. Speciﬁcally, it counts the number of times each letter appears as the ﬁrst letter in a word, as the last letter in a word, and as an interior letter in a word (i.e. a letter that is neither ﬁrst nor last).